Refactor ViewportDataBinary to be a readonly struct

+++ GameEngine/Util/CoordinateSystem.cs
@@ -53,10 +53,10 @@ public static class CoordinateSystem
         return (chunkX * Config.CHUNK_SIZE + localX, chunkY * Config.CHUNK_SIZE + localY);
     }
 
-    public static (int ChunkX, int ChunkY, byte LocalX, byte LocalY) WorldToChunkAndLocal(int worldX, int worldY)
+    public static (byte ChunkX, byte ChunkY, byte LocalX, byte LocalY) WorldToChunkAndLocal(int worldX, int worldY)
     {
-        int chunkX = worldX / Config.CHUNK_SIZE;
-        int chunkY = worldY / Config.CHUNK_SIZE;
+        byte chunkX = worldX / Config.CHUNK_SIZE;
+        byte chunkY = worldY / Config.CHUNK_SIZE;
         byte localX = (byte)(worldX % Config.CHUNK_SIZE);
         byte localY = (byte)(worldY % Config.CHUNK_SIZE);
         return (chunkX, chunkY, localX, localY);
